pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "build"]
build-backend = "setuptools.build_meta"

[project]
name = "fellowai"
version = "0.1.0"
description = "An agentic application for reading and summarizing research papers."
readme = "README.md"
requires-python = ">=3.8"
authors = [
    {name = "Brad", email = "brad@example.com"}
]
dependencies = [
    "pydantic>=2.0.0",
    "crewai[litellm]",
    "langgraph",
    "langchain-openai",
    "python-dotenv"
]

[project.optional-dependencies]
dev = [
    "pytest",
    "build",
]

[tool.setuptools.packages.find]
where = ["src"]
